backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
starlette>=0.36.3,<0.37.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.9
celery[redis]==5.3.6
redis==5.0.1
httpx>=0.27.0
beautifulsoup4==4.12.3
playwright==1.41.1
openrouter>=0.3.0
python-dotenv==1.0.1
feedparser==6.0.11
python-dateutil==2.9.0
email-validator==2.1.1
requests==2.31.0
sse-starlette==1.8.2
